Fixes an issue with not being able to write default settings file and remove same.

+++ common_tools/immunogrep_igfft_functions.py
@@ -1495,7 +1495,8 @@ def GetDefaultParameters():
 			'times_above_ratio': None,
 			's_fft': None						
 		}
-	os.remove('defaultsettings_fftprogram.txt')
+	if os.path.isfile('defaultsettings_fftprogram.txt')
+		os.remove('defaultsettings_fftprogram.txt')
 
 	return default_settings
